OPENEUROPA-1383: Move method about assert rendering to Trait.
